refactor(ioxide): group the Protocol folder by protocol family
Fourteen files sat flat in Protocol/, and nothing in the listing said which
belonged to which protocol. The dependency graph already answered it: the six
response-writing files are reachable only from Http1Driver, and the six
Multiplexed ones only from Http2Driver and Http3Driver.

    Protocol/
      ConnectionDriver.cs   the TCP entry point, and the only fork between them
      Http1/                Http1Driver + its response writing and sinks
      Multiplexed/          Http2Driver, Http3Driver + what the two share

Namespaces follow the folders, so the moved types are now under .Protocol.Http1
and .Protocol.Multiplexed. Both nest inside .Protocol, which is how the drivers
still reach ConnectionDriver without importing anything.
