+++ Engine/Ioxide/Protocol/Http1/DateHeader.cs
@@ -1,6 +1,6 @@
 using System.Runtime.CompilerServices;
 
-namespace GenHTTP.Engine.Ioxide.Protocol;
+namespace GenHTTP.Engine.Ioxide.Protocol.Http1;
 
 /// <summary>
 /// Per-reactor cached "Date: ...\r\n" header, refreshed at most once a second. [ThreadStatic]
